#37 [REFACTOR] 소셜링크를 컴포넌트로 분리
