+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { default as PageLayout } from './PageLayout';
 export { default as Header } from './Header';
 export { default as Footer } from './Footer';
+export { default as SocialMediaLinkCards } from './SocialMediaLinkCards';
